Add credit card fields into formRegisterUser

+++ account/views.py
@@ -24,7 +24,7 @@ import base64
 import requests
 import json
 from ezonseller import settings
-
+from account.service import create_card, serialize_credit_card
 
 class Login(APIView):
     """
@@ -114,6 +114,8 @@ class RegisterView(APIView):
     permission_classes = (permissions.AllowAny,)
 
     def post(self, request):
+        user = request.data.get('user')
+
         if not request.data.get('callback'):
             return Response({"message": "reCAPTCHA field cant not be empty"}, status=status.HTTP_400_BAD_REQUEST)
         recaptcha_response = request.data['callback']
@@ -123,7 +125,9 @@ class RegisterView(APIView):
         })
         if not json.loads(r.content.decode())['success']:
             return Response({'message': 'Invalid reCAPTCHA. Please try again.'}, status=status.HTTP_400_BAD_REQUEST)
-        user_serializer = validations.UserCreateSerializers(data=request.data)
+
+        user_serializer = validations.UserCreateSerializers(data=user)
+
         if user_serializer.is_valid() is False:
             errors_msg = []
             errors_keys = list(user_serializer.errors.keys())
@@ -131,10 +135,25 @@ class RegisterView(APIView):
                 errors_msg.append(str(i) + ": " + str(user_serializer.errors[i][0]))
             error_msg = "".join(errors_msg)
             return Response({'message': errors_msg[0]}, status=status.HTTP_400_BAD_REQUEST)
+
         user = user_serializer.save()
+
+        serialize_card = serialize_credit_card(request,user)
+
+        if 'message' in serialize_card:
+            error_serialize_card = serialize_card['message']
+            return Response({'message': error_serialize_card}, status=status.HTTP_400_BAD_REQUEST)
+
+        created_card  = create_card(serialize_card,user)
+
+        if 'message' in created_card:
+            error_create_card = created_card['message']
+            return Response({'message': error_create_card}, status=status.HTTP_400_BAD_REQUEST)
+
+
         if notify_views.activate_account(user, request):
             return Response({'username': user.username, 'message': 'The email has been send'},
-                            status=status.HTTP_201_CREATED)
+                        status=status.HTTP_201_CREATED)
         else:
             user.delete()
             return Response({'message': 'The email cannot be sent and account not created'},
